Backend - Move updatePassword & deleteAccount to users route
