+++ backend/routes/auth.js
@@ -327,76 +327,6 @@ router.put('/:user_id', checkUserLogged, storage.upload.single('picture'), updat
 })
 
 
-const updatePassword = async(request, response, next) => {
-    try {
-        const newPassword = processInput(request.body.newPassword, 'hardVC', 'user password');
-        const confirmPassword = processInput(request.body.confirmPassword, 'hardVC', 'user password');
-        const targetId = processInput(request.params.user_id, 'idNum', 'user id');
-        const loggedUserId = request.user.a_id || request.user.v_id || request.user.f_id;
-
-        if (targetId === loggedUserId && newPassword === confirmPassword) {
-            const loggedUserEmail = request.user.a_email || request.user.v_email || request.user.f_email;
-            const hashedPassword = await hashPassword(newPassword);
-            
-            await usersQueries.updatePassword(loggedUserEmail, hashedPassword);
-            next();
-        } else {
-            throw new Error("401__Not authorized to update - password doesn't match or you don't have the right to update");
-        }
-
-    } catch (err) {
-        handleError(err, request, response, next);
-    }
-}
-
-router.patch('/:user_id', checkUserLogged, updatePassword, (request, response) => {
-    response.json({
-        error: false,
-        message: 'Successfully updated password',
-        payload: request.user
-    })
-})
-
-
-const deleteAccount = async(request, response, next) => {
-    try {
-        const loggedUserId = request.user.a_id || request.user.v_id || request.user.f_id;
-        const targetId = processInput(request.params.user_id, 'idNum', 'user id');
-        const loggedUserEmail = request.user.a_email || request.user.v_email || request.user.f_email;
-
-        if (targetId === loggedUserId) {
-            if (request.user.a_id) {
-                await adminQueries.deleteAdmin(loggedUserId);
-            } 
-            else if (request.user.v_id) {
-                await volunteersQueries.deleteVolunteer(loggedUserId);
-            } 
-            else {
-                await fellowsQueries.deleteFellow(loggedUserId);
-            }
-    
-            // await usersQueries.deleteUser(loggedUserEmail);
-            next();
-        } 
-        else {
-            throw new Error("401__Not authorized to delete");
-        }
-
-    } catch (err) {
-        handleError(err, request, response, next);
-    }
-}
-
-router.delete('/:user_id', checkUserLogged, deleteAccount, (request, response) => {
-    request.logOut();
-    response.json({
-        error: false,
-        message: 'Successfully deleted user/brand',
-        payload: null
-    })
-})
-
-
 router.get('/logout', checkUserLogged, (request, response) => {
     request.logOut();
     response.json({
